ingestion/requirements.txt
fastapi
uvicorn
pydantic
qdrant-client
google-genai
unstructured-client
FlagEmbedding
transformers<5.0.0
torch
